moved one_shot_object_detection inside object_detection
