+++ src/unity/toolkits/object_detection/one_shot_object_detection/class_registrations.hpp
